refactor: node package export and deps
