+++ Cargo.toml
@@ -125,7 +125,7 @@ tokenizers = "0.21.1"
 sentencepiece = "0.11.3"
 
 [workspace.package]
-version = "0.8.0"
+version = "0.8.1"
 edition = "2024"
 rust-version = "1.86"
 exclude = [".github/"]
