+++ sdks/node/npm/darwin-arm64/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@lazai-labs/alith-darwin-arm64",
   "repository": "https://github.com/0xLazAI/alith",
-  "version": "0.8.0",
+  "version": "0.8.1",
   "os": [
     "darwin"
   ],
