+++ sdks/node/npm/linux-x64-gnu/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@lazai-labs/alith-linux-x64-gnu",
   "repository": "https://github.com/0xLazAI/alith",
-  "version": "0.8.0",
+  "version": "0.8.1",
   "os": [
     "linux"
   ],
