+++ sdks/node/npm/win32-arm64-msvc/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@lazai-labs/alith-win32-arm64-msvc",
   "repository": "https://github.com/0xLazAI/alith",
-  "version": "0.8.0",
+  "version": "0.8.1",
   "os": [
     "win32"
   ],
