+++ sdks/node/package.json
@@ -1,6 +1,6 @@
 {
   "name": "alith",
-  "version": "0.8.0",
+  "version": "0.8.1",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/0xLazAI/alith.git"
@@ -8,6 +8,33 @@
   "license": "Apache-2.0",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "dist/index.js",
+      "require": "dist/index.js",
+      "types": "dist/index.d.ts"
+    },
+    "./data": {
+      "import": "dist/data/index.js",
+      "require": "dist/data/index.js",
+      "types": "dist/data/index.d.ts"
+    },
+    "./data/storage": {
+      "import": "dist/data/storage/index.js",
+      "require": "dist/data/storage/index.js",
+      "types": "dist/data/storage/index.d.ts"
+    },
+    "./lazai": {
+      "import": "dist/lazai/index.js",
+      "require": "dist/lazai/index.js",
+      "types": "dist/lazai/index.d.ts"
+    },
+    "./lazai/node": {
+      "import": "dist/lazai/node/index.js",
+      "require": "dist/lazai/node/index.js",
+      "types": "dist/lazai/node/index.d.ts"
+    }
+  },
   "napi": {
     "name": "alith",
     "package": {
@@ -86,9 +113,6 @@
     "node-rsa": "^1.1.1",
     "openpgp": "^6.1.1",
     "web3": "^4.16.0",
-    "web3-eth-accounts": "^4.3.1",
-    "web3-eth-contract": "^4.7.2",
-    "web3-utils": "^4.3.3",
     "zod": "3.24.4",
     "zod-to-json-schema": "^3.24.1"
   }
