+++ sdks/node/src/extractor.ts
